+++ src/components/element-configurator/Container/layout-icons/FullThree.vue
