+++ src/components/element-configurator/Container/layout-icons/OneTwo.vue
