+++ src/components/element-configurator/Content/TextAlign.vue
