+++ src/components/element-configurator/ElementConfigurator.vue
@@ -2,7 +2,11 @@
   <div class="p-4 mt-4 border-t border-gray-200">
     <h2 class="w-full text-center text-gray-600 text-lg font-bold">Element Configuration</h2>
     <div v-if="activeComponent.length > 0" class="flex flex-col">
-      <component v-for="(component, i) in activeComponent" :is="component.name" :configName="component.configName" :key="i"></component>
+      <component v-for="(component, i) in activeComponent"
+        :key="i"
+        :is="component.name"
+        :configName="component.configName">
+      </component>
     </div>
     <div class="mt-4 py-8 bg-gray-100 text-center text-gray-400 italic" v-else>Select an element first</div>
   </div>
@@ -13,17 +17,17 @@ import { computed } from 'vue'
 import { useStore } from 'vuex'
 
 // Content element components
-import ContentFontColor from './ContentFontColor.vue'
-import ContentTextAlign from './ContentTextAlign.vue'
-import ContentHeadingLevel from './ContentHeadingLevel.vue'
+import ContentFontColour from './Content/FontColour.vue'
+import ContentTextAlign from './Content/TextAlign.vue'
+import ContentHeadingLevel from './Content/HeadingLevel.vue'
 
 // Container element components
-import ContainerLayoutMode from './ContainerLayoutMode.vue'
+import ContainerLayoutMode from './Container/LayoutMode.vue'
 
 export default {
   components: {
     // Content element components
-    "content-font-color": ContentFontColor,
+    "content-font-colour": ContentFontColour,
     "content-text-align": ContentTextAlign,
     "content-heading-level": ContentHeadingLevel,
 
@@ -39,11 +43,11 @@ export default {
         case 'heading':
           result = [
             {name: "content-heading-level", configName: "headingLevel"},
-            {name: "content-font-color", configName: "fontColour"},
+            {name: "content-font-colour", configName: "fontColour"},
             {name: "content-text-align", configName: "textAlign"}
           ]
           break;
-        case 'container':
+        case 'division':
           result = [{ name: "container-layout-mode", configName: "layoutMode"}]
           break;
       }
